Create cube root calculator
this code here asks the user a number and then asks the user again that what does the user thinks the cube root of the number may be and then uses the "Newton-Raphson" formula which  is a technique used to find approximate solutions to equations. It starts with a guess and keeps improving it using a formula based on the function and its derivative, until the result is close enough to the actual solution and then gives the approximate ans.